+++ crucible/src/Lang/Crucible/CFG/ExtractSubgraph.hs
@@ -44,7 +44,7 @@ extractSubgraph :: (KnownCtx TypeRepr init, KnownRepr TypeRepr ret)
                 -> BlockID blocks init
                 -> HandleAllocator
                 -> IO (Maybe (SomeCFG ext init ret))
-extractSubgraph (CFG{cfgBlockMap = orig, cfgBreakpoints = breakpoints}) cuts bi halloc =
+extractSubgraph (CFG{cfgBlockMap = orig, cfgCutpoints = cutpoints}) cuts bi halloc =
   extractSubgraphFirst orig cuts MapF.empty zeroSize bi $
     \(SubgraphIntermediate finalMap finalInitMap _sz entryID cb) -> do
         hn <- mkHandle halloc startFunctionName
@@ -54,9 +54,9 @@ extractSubgraph (CFG{cfgBlockMap = orig, cfgBreakpoints = breakpoints}) cuts bi 
             { cfgBlockMap = bm
             , cfgEntryBlockID = entryID
             , cfgHandle = hn
-            , cfgBreakpoints = Bimap.fromList $ Map.toList $
+            , cfgCutpoints = Bimap.fromList $ Map.toList $
                 Map.mapMaybe (viewSome $ \bid -> Some <$> MapF.lookup bid finalMap) $
-                Bimap.toMap breakpoints
+                Bimap.toMap cutpoints
             }
 
 -- | Type for carrying intermediate results through subraph extraction